refactor: Refactored the code in App.js to make it more concise keeping in mind the component based design methodology.
The application currently has only three shelves. But going forward
there is possibility that more shelves will be added. So each time a new
shelf is added, the code in App.js needs to be modfied. But with
refactoring done in this commit, new shleves can be added by simply
adding a new entry into an array of shelves and thats it.

Shelves array looks like this
[ { id: 'uniqueShelfID', name: 'A identifiable shelf name' }, ... ]