refactor: Converted BookShelf component from stateless functional component to state component and also refactored Book component to support multiple authors.
The BookShelf had to be converted so it could support displaying a
simple spinner, thus improving the UI experience. The spinner is a pure
CSS class which is displayed based on the loading state of the component.
This state is directly linked to the fact that if the books are loaded
or not.

The author props of Book component had to be changed to authors so as
to display multiple authors of a book instead of assuming one author.
I realised this only after viewing the json response of BooksAPI.getAll().