feat: Added a new Components SearchBar for handling search query and accordingly modify the books list.
Created a separate component, because in ListSearchBooks
it was becoming difficult to control rendering of list of
books with two states QUERY & BOOKS. For example, since
I am calling search API in componentDidUpdate, so in case user
clears content of search box, I should not make the API call
with empty string and also avoid updating component. But then
the list books will not be cleared and /search will display books
even though there is no query in search box.

Also if the Search API is returning an error there is no
need to update the displayed list of books on /search.