refactor: Changes made in App compoonent to only update list of books correctly. Also ListBooks is a functional stateless component.
As there is only one list of books so just by filtering out
all books which are not be updated and add the book to be updated
to this new filtere list can resolve our problem.

ListBooks had to be made stateless as shelves are never
changing so can be defined in a constant.